// MuvClips/src/app/models/movie/movie.ts
export class Movie {
    id?: string;
    title: string;
    description: string;
    duration: string;
    actors: string[];
    directors: string[];
    genre: string;
    classification: string;
    lenguage: string;
    punctuation: string;
    new: boolean;
    cover: string;
    movie: string;
    type: string;
}

/*
titulo
descripcion
duracion
actores
directores
genero
clasificaciè´¸n
idioma
calificacion
estreno true or false
caratula
pelicula
*/